DO NOT MERGE: libcameraservice: Fix nullptr crash when no client.
This change prevents a crash in the camera service when the camera
HAL notifies the service about the absence of a removable camera and
there happens to be no client connected to the service. It checks
that the pointer returned from clientToDisconnect.get() is non-null
before trying to dereference it (as is done in existing code
immediately below this change).

Bug: 25165323
Change-Id: I8055654bac980542e63ea7f52bf897eaafbc09bc
Signed-off-by: Tom Keel <thomas.keel@intel.com>
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>

 Conflicts:
	services/camera/libcameraservice/CameraService.cpp